==> 01_ETL.ipynb <==
import os
import pandas as pd

RAW_DIR = r"C:\Users\12432\Desktop\archive"
PROCESSED_DIR = "./data/processed"
os.makedirs(PROCESSED_DIR, exist_ok=True)

orders = pd.read_csv(
    RAW_DIR + r"\olist_orders_dataset.csv",
    low_memory=False,
    parse_dates=[
        "order_purchase_timestamp",
        "order_approved_at",
        "order_delivered_carrier_date",
        "order_delivered_customer_date",
        "order_estimated_delivery_date",
    ],
)
customers = pd.read_csv(RAW_DIR + r"\olist_customers_dataset.csv", low_memory=False)
items     = pd.read_csv(RAW_DIR + r"\olist_order_items_dataset.csv",
                        low_memory=False, parse_dates=["shipping_limit_date"])
payments  = pd.read_csv(RAW_DIR + r"\olist_order_payments_dataset.csv", low_memory=False)
reviews   = pd.read_csv(RAW_DIR + r"\olist_order_reviews_dataset.csv",
                        low_memory=False, parse_dates=["review_creation_date", "review_answer_timestamp"])
products  = pd.read_csv(RAW_DIR + r"\olist_products_dataset.csv", low_memory=False)
sellers   = pd.read_csv(RAW_DIR + r"\olist_sellers_dataset.csv", low_memory=False)
geo       = pd.read_csv(RAW_DIR + r"\olist_geolocation_dataset.csv", low_memory=False)
catmap    = pd.read_csv(RAW_DIR + r"\product_category_name_translation.csv", low_memory=False)



orders["actual_delivery_days"] = (
    orders["order_delivered_customer_date"] - orders["order_purchase_timestamp"]
).dt.days

orders["delivery_delay_days"] = (
    orders["order_delivered_customer_date"] - orders["order_estimated_delivery_date"]
).dt.days

orders["is_late"] = orders["delivery_delay_days"] > 0

orders[["actual_delivery_days","delivery_delay_days","is_late"]].describe(include="all")


orders.to_parquet(PROCESSED_DIR + "/orders.parquet", index=False)
customers.to_parquet(PROCESSED_DIR + "/customers.parquet", index=False)
items.to_parquet(PROCESSED_DIR + "/order_items.parquet", index=False)
payments.to_parquet(PROCESSED_DIR + "/order_payments.parquet", index=False)
reviews.to_parquet(PROCESSED_DIR + "/order_reviews.parquet", index=False)
products.to_parquet(PROCESSED_DIR + "/products.parquet", index=False)
sellers.to_parquet(PROCESSED_DIR + "/sellers.parquet", index=False)
geo.to_parquet(PROCESSED_DIR + "/geolocation.parquet", index=False)
catmap.to_parquet(PROCESSED_DIR + "/category_translation.parquet", index=False)


orders_chk = pd.read_parquet(PROCESSED_DIR + "/orders.parquet")
items_chk  = pd.read_parquet(PROCESSED_DIR + "/order_items.parquet")

print("orders_chk:", orders_chk.shape, "items_chk:", items_chk.shape)
orders_chk[["order_purchase_timestamp","order_delivered_customer_date","actual_delivery_days"]].head(3)
